+++ CHANGELOG.md
@@ -1,3 +1,176 @@
+## [1.114.5](https://github.com/adobe/spacecat-api-service/compare/v1.114.4...v1.114.5) (2025-05-26)
+
+
+### Bug Fixes
+
+* **deps:** update adobe fixes ([#931](https://github.com/adobe/spacecat-api-service/issues/931)) ([03b0913](https://github.com/adobe/spacecat-api-service/commit/03b0913a63536a1a1d4ce39ff3c5ebfdd56daa7f))
+
+## [1.114.4](https://github.com/adobe/spacecat-api-service/compare/v1.114.3...v1.114.4) (2025-05-25)
+
+
+### Bug Fixes
+
+* **deps:** update adobe fixes ([#930](https://github.com/adobe/spacecat-api-service/issues/930)) ([224e4fe](https://github.com/adobe/spacecat-api-service/commit/224e4fe3fc52717ca54311aadca25c70044647de))
+
+## [1.114.3](https://github.com/adobe/spacecat-api-service/compare/v1.114.2...v1.114.3) (2025-05-24)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#929](https://github.com/adobe/spacecat-api-service/issues/929)) ([8ab9bb2](https://github.com/adobe/spacecat-api-service/commit/8ab9bb23551bd64a1033b1c594aa32380db327b0))
+
+## [1.114.2](https://github.com/adobe/spacecat-api-service/compare/v1.114.1...v1.114.2) (2025-05-23)
+
+
+### Bug Fixes
+
+* **deps:** update adobe fixes ([#928](https://github.com/adobe/spacecat-api-service/issues/928)) ([c787f38](https://github.com/adobe/spacecat-api-service/commit/c787f387afce45f57b948431b685c7782e3daf8d))
+
+## [1.114.1](https://github.com/adobe/spacecat-api-service/compare/v1.114.0...v1.114.1) (2025-05-22)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/spacecat-shared-data-access to v2.19.4 ([#927](https://github.com/adobe/spacecat-api-service/issues/927)) ([e0d5bb0](https://github.com/adobe/spacecat-api-service/commit/e0d5bb0487b923c4a837b6ea2d2e77aa131a1c7b))
+
+# [1.114.0](https://github.com/adobe/spacecat-api-service/compare/v1.113.0...v1.114.0) (2025-05-22)
+
+
+### Features
+
+* mcp prototype server ([#923](https://github.com/adobe/spacecat-api-service/issues/923)) ([6a2c924](https://github.com/adobe/spacecat-api-service/commit/6a2c924744a4f9d31905d91e9f3be24b85def175))
+
+# [1.113.0](https://github.com/adobe/spacecat-api-service/compare/v1.112.0...v1.113.0) (2025-05-22)
+
+
+### Features
+
+* [Preflight] Split identify and suggest flows - a small fix ([#926](https://github.com/adobe/spacecat-api-service/issues/926)) ([c763e09](https://github.com/adobe/spacecat-api-service/commit/c763e09c30a5e64654f160cae56c182ed523d58a))
+
+# [1.112.0](https://github.com/adobe/spacecat-api-service/compare/v1.111.0...v1.112.0) (2025-05-22)
+
+
+### Features
+
+* get files by key from s3 ([#921](https://github.com/adobe/spacecat-api-service/issues/921)) ([342b4c5](https://github.com/adobe/spacecat-api-service/commit/342b4c54feceb0f72ae4abccbc47e161c44e9f59))
+
+# [1.111.0](https://github.com/adobe/spacecat-api-service/compare/v1.110.2...v1.111.0) (2025-05-22)
+
+
+### Features
+
+* [Preflight] Split identify and suggest flows ([#924](https://github.com/adobe/spacecat-api-service/issues/924)) ([68dfffa](https://github.com/adobe/spacecat-api-service/commit/68dfffa87977cdd4dc5cd2c9d4cb54933334815b))
+
+## [1.110.2](https://github.com/adobe/spacecat-api-service/compare/v1.110.1...v1.110.2) (2025-05-21)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/spacecat-shared-http-utils to v1.13.2 ([#925](https://github.com/adobe/spacecat-api-service/issues/925)) ([0086bca](https://github.com/adobe/spacecat-api-service/commit/0086bca07ea5fbdf136c3595b89ce96ded32882e))
+
+## [1.110.1](https://github.com/adobe/spacecat-api-service/compare/v1.110.0...v1.110.1) (2025-05-20)
+
+
+### Bug Fixes
+
+* **deps:** update adobe fixes ([#922](https://github.com/adobe/spacecat-api-service/issues/922)) ([af1e944](https://github.com/adobe/spacecat-api-service/commit/af1e94491de792fcbd0c15d0e17ca5bbf6e58fb5))
+
+# [1.110.0](https://github.com/adobe/spacecat-api-service/compare/v1.109.4...v1.110.0) (2025-05-20)
+
+
+### Bug Fixes
+
+* **deps:** update adobe fixes ([#920](https://github.com/adobe/spacecat-api-service/issues/920)) ([5ac168e](https://github.com/adobe/spacecat-api-service/commit/5ac168e871dd14e290f1d23161e9b78351eaf7db))
+
+
+### Features
+
+* add scraper content retrieval ([#913](https://github.com/adobe/spacecat-api-service/issues/913)) ([55e4700](https://github.com/adobe/spacecat-api-service/commit/55e470057995e05e4a4c4b9fbb9c61547f19eddc))
+
+## [1.109.5](https://github.com/adobe/spacecat-api-service/compare/v1.109.4...v1.109.5) (2025-05-20)
+
+
+### Bug Fixes
+
+* **deps:** update adobe fixes ([#920](https://github.com/adobe/spacecat-api-service/issues/920)) ([5ac168e](https://github.com/adobe/spacecat-api-service/commit/5ac168e871dd14e290f1d23161e9b78351eaf7db))
+
+## [1.109.4](https://github.com/adobe/spacecat-api-service/compare/v1.109.3...v1.109.4) (2025-05-17)
+
+
+### Bug Fixes
+
+* **deps:** update adobe fixes ([#917](https://github.com/adobe/spacecat-api-service/issues/917)) ([ca4e313](https://github.com/adobe/spacecat-api-service/commit/ca4e313eca8132280c59eb38e0e36a34968a24b4))
+
+## [1.109.3](https://github.com/adobe/spacecat-api-service/compare/v1.109.2...v1.109.3) (2025-05-17)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#916](https://github.com/adobe/spacecat-api-service/issues/916)) ([ff4367c](https://github.com/adobe/spacecat-api-service/commit/ff4367ce99df5b1e56bd564a0acefff6a38a3be5))
+
+## [1.109.2](https://github.com/adobe/spacecat-api-service/compare/v1.109.1...v1.109.2) (2025-05-16)
+
+
+### Bug Fixes
+
+* **deps:** update adobe fixes ([#915](https://github.com/adobe/spacecat-api-service/issues/915)) ([30ff614](https://github.com/adobe/spacecat-api-service/commit/30ff6148c1c06042dfef0202d2830532f3770f50))
+
+## [1.109.1](https://github.com/adobe/spacecat-api-service/compare/v1.109.0...v1.109.1) (2025-05-15)
+
+
+### Bug Fixes
+
+* **deps:** update adobe fixes ([#914](https://github.com/adobe/spacecat-api-service/issues/914)) ([32ea20e](https://github.com/adobe/spacecat-api-service/commit/32ea20ee2fed94ffedb7fa5e6d65abafd92c7a22))
+
+# [1.109.0](https://github.com/adobe/spacecat-api-service/compare/v1.108.1...v1.109.0) (2025-05-15)
+
+
+### Features
+
+* Slack command for internal reports ([#901](https://github.com/adobe/spacecat-api-service/issues/901)) ([89951e2](https://github.com/adobe/spacecat-api-service/commit/89951e28b752ec5c602ab3ed9dcdcd053c278e44))
+
+## [1.108.1](https://github.com/adobe/spacecat-api-service/compare/v1.108.0...v1.108.1) (2025-05-15)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/spacecat-shared-utils to v1.38.0 ([#912](https://github.com/adobe/spacecat-api-service/issues/912)) ([31ed040](https://github.com/adobe/spacecat-api-service/commit/31ed0408fc58956cd7b028ae4d6a8b77ab6c2a5b))
+
+# [1.108.0](https://github.com/adobe/spacecat-api-service/compare/v1.107.12...v1.108.0) (2025-05-14)
+
+
+### Features
+
+* implement preflight job creation and send SQS notification from preflight endpoint ([#902](https://github.com/adobe/spacecat-api-service/issues/902)) ([88b0c64](https://github.com/adobe/spacecat-api-service/commit/88b0c641e08463df4fc4f381138cc7b7758cf49c))
+
+## [1.107.12](https://github.com/adobe/spacecat-api-service/compare/v1.107.11...v1.107.12) (2025-05-14)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/spacecat-shared-data-access to v2.19.0 ([#911](https://github.com/adobe/spacecat-api-service/issues/911)) ([f66d838](https://github.com/adobe/spacecat-api-service/commit/f66d83860bc4a2f7516b0c2c5b5854b18f5d8f6e))
+
+## [1.107.11](https://github.com/adobe/spacecat-api-service/compare/v1.107.10...v1.107.11) (2025-05-14)
+
+
+### Bug Fixes
+
+* handle anonymous site detection endpoints in access control utils ([#910](https://github.com/adobe/spacecat-api-service/issues/910)) ([726cffd](https://github.com/adobe/spacecat-api-service/commit/726cffdb6b0e17bb9e36b47c9ff2fc7b4daa437c))
+
+## [1.107.10](https://github.com/adobe/spacecat-api-service/compare/v1.107.9...v1.107.10) (2025-05-12)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/spacecat-shared-http-utils to v1.11.0 ([#908](https://github.com/adobe/spacecat-api-service/issues/908)) ([c1ac706](https://github.com/adobe/spacecat-api-service/commit/c1ac70618f92cf8cd93cc9f72fa1ad8e7037ce81))
+
+## [1.107.9](https://github.com/adobe/spacecat-api-service/compare/v1.107.8...v1.107.9) (2025-05-12)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/spacecat-shared-data-access to v2.18.0 ([#905](https://github.com/adobe/spacecat-api-service/issues/905)) ([a1faf90](https://github.com/adobe/spacecat-api-service/commit/a1faf90d1fe4c7ca18f9cc0e299873dbc4f812de))
+
 ## [1.107.8](https://github.com/adobe/spacecat-api-service/compare/v1.107.7...v1.107.8) (2025-05-11)